refactor: moved file locations on lib folder.
