+++ src/lib/get-wallet-balance.js
