+++ src/lib/local-storage.js
@@ -5,19 +5,9 @@ const localStorage = {
   DEGENICS_KS: 'DEGENICS_KS',
   DEGENICS_RPC_URL: 'DEGENICS_RPC_URL',
   DEGENICS_CONFIG: 'DEGENICS_CONFIG',
-  DEBIO_WALLET: 'DEBIO_WALLET',
   /**
    * getter setters
    */
-  setWallet(wallet) {
-    window.localStorage.setItem(this.DEBIO_WALLET, wallet)
-  },
-  getWallet() {
-    return window.localStorage.getItem(this.DEBIO_WALLET)
-  },
-  clearWallet() {
-    window.localStorage.removeItem(this.DEBIO_WALLET)
-  },
   setKeystore(keystore) {
     window.localStorage.setItem(this.DEGENICS_KS, keystore)
   },
@@ -43,7 +33,6 @@ const localStorage = {
     window.localStorage.removeItem(this.DEGENICS_KS)
     window.localStorage.removeItem(this.DEGENICS_RPC_URL)
     window.localStorage.removeItem(this.DEGENICS_CONFIG)
-    window.localStorage.removeItem(this.DEBIO_WALLET)
   }
 }
