+++ src/lib/send-transaction.js
@@ -1,5 +1,5 @@
-import store from '../../store'
-import appConfig from '../app-config'
+import store from '../store'
+import appConfig from './app-config'
 import * as EthUtil from 'ethereumjs-util'
 import Transaction from 'ethereumjs-tx'
