+++ src/lib/specimen-files-temp-store.js
