+++ src/lib/string-format.js
