+++ src/plugins/vuetify.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import Vuetify from 'vuetify/lib/framework';
+import Vuetify from 'vuetify';
 import SkinIcon from '../components/icons/SkinIcon'
 
 Vue.use(Vuetify);
